+++ charmcraft.yaml
@@ -42,24 +42,9 @@ config:
 assumes:
   - k8s-api
 
-bases:
-  - build-on:
-    - name: ubuntu
-      channel: "22.04"
-      architectures: ["amd64"]
-    run-on:
-    - name: ubuntu
-      channel: "22.04"
-      architectures: ["amd64"]
-  - build-on:
-    - name: ubuntu
-      channel: "22.04"
-      architectures: ["arm64"]
-    run-on:
-    - name: ubuntu
-      channel: "22.04"
-      architectures: ["arm64"]
-
+platforms:
+  ubuntu@24.04:amd64:
+  ubuntu@24.04:arm64:
 
 
 containers:
@@ -98,10 +83,13 @@ requires:
 
 parts:
   charm:
-    plugin: charm
-    # include rustc and cargo to compile pydantic
-    build-packages: [git, rustc, cargo]
-    charm-binary-python-packages: ["pydantic>=2", "cryptography", "jsonschema", "opentelemetry-exporter-otlp-proto-http==1.21.0"]
+    source: .
+    plugin: uv
+    build-packages: [git]
+    build-snape: [astral-uv]
+    override-build: |
+      craftctl default
+      git describe --always > $CRAFT_PART_INSTALL/version
   istioctl:
     plugin: dump
     source: https://github.com/istio/istio/releases/download/1.24.0/istioctl-1.24.0-linux-amd64.tar.gz
