+++ pyproject.toml
@@ -1,6 +1,34 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+[project]
+name = "istio-k8s"
+version = "0.0"
+requires-python = "~=3.8"
+
+dependencies = [
+    "ops~=2.5",
+    "pydantic>=2",
+    "cosl",
+    "opentelemetry-exporter-otlp-proto-http==1.21.0",
+    "lightkube-extensions@git+https://github.com/canonical/lightkube-extensions.git@main",
+]
+
+[project.optional-dependencies]
+dev = [
+    "ruff",
+    "codespell",
+    "pyright",
+    "pytest",
+    "coverage[toml]",
+    "pytest-mock",
+    "ops[testing]",
+    "pytest",
+    "pytest-asyncio==0.21.2",
+    "juju",
+    "pytest-operator",
+]
+
 # Testing tools configuration
 [tool.coverage.run]
 branch = true
@@ -12,6 +40,7 @@ show_missing = true
 minversion = "6.0"
 log_cli_level = "INFO"
 asyncio_mode = "auto"
+addopts = "--tb=native --verbose --capture=no --log-cli-level=INFO"
 
 # Formatting tools configuration
 [tool.black]
@@ -41,8 +70,12 @@ extend-ignore = [
 ignore = ["E501", "D107"]
 per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104"]}
 
+[toos.ruff.lint.pydocstyle]
+convention = "google"
+
 [tool.codespell]
 skip = "build,lib,venv,icon.svg,.tox,.git,.mypy_cache,.ruff_cache,.coverage"
 
 [tool.pyright]
+extraPaths = ["src", "lib"]
 include = ["src/**.py"]
