+++ tests/unit/test_istio_metadata_lib.py
