+++ tox.ini
@@ -3,83 +3,58 @@
 
 [tox]
 skip_missing_interpreters = True
-envlist = lint, static-{charm}, unit, scenario, integration
+envlist = lint, static, unit
 
 [vars]
 src_path = {toxinidir}/src
 lib_path = {toxinidir}/lib/charms/istio_k8s
 tst_path = {toxinidir}/tests
 all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
+uv_flags = --frozen --isolated --extra=dev
 
 [testenv]
+allowlist_externals = uv
 basepython = python3
 setenv =
   PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
   PYTHONBREAKPOINT=ipdb.set_trace
   PY_COLORS=1
 
+[testenv:lock]
+description = Update uv.lock with the latest deps
+commands =
+  uv lock --upgrade --no-cache
+
 [testenv:fmt]
 description = Apply coding style standards to code
-deps =
-    black
-    ruff
 commands =
-    ruff check --fix {[vars]all_path}
-    black {[vars]all_path}
+  uv run {[vars]uv_flags} ruff check --fix-only {[vars]all_path}
 
 [testenv:lint]
 description = Check code against coding style standards
-deps =
-    black
-    ruff
-    codespell
 commands =
-    codespell {toxinidir} --skip {toxinidir}/src/manifests
-    ruff check {[vars]all_path}
-    black --check --diff {[vars]all_path}
+  uv run {[vars]uv_flags} ruff check {[vars]all_path}
 
-[testenv:static-{charm,lib}]
+[testenv:static]
 skip_install=True
 description = Run static analysis checks
-deps =
-    pyright
-    -r {toxinidir}/requirements.txt
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
 commands =
-    charm: pyright {[vars]src_path}
-    lib:
+  uv run {[vars]uv_flags} pyright {[vars]src_path} {[vars]lib_path}
+  /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
 
 [testenv:unit]
 description = Run unit tests
-deps =
-    pytest
-    pytest-mock
-    coverage[toml]
-    -r {toxinidir}/requirements.txt
 commands =
-    coverage run \
-      --source={[vars]src_path} \
-      -m pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/unit
-    coverage report -m
-
-[testenv:scenario]
-description = Run scenario tests
-deps =
-    -r{toxinidir}/requirements.txt
-    pytest
-    ops[testing]
-commands =
-    pytest -vv --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/scenario
+  uv run {[vars]uv_flags} coverage run --source={[vars]src_path} -m pytest {[vars]tst_path}/unit {posargs}
+  uv run {[vars]uv_flags} coverage report
 
 [testenv:integration]
 description = Run integration tests
-deps =
-    pytest
-    pytest-asyncio==0.21.2
-    juju
-    pytest-operator
-    -r {toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/integration
+  uv run {[vars]uv_flags} pytest --exitfirst {[vars]tests_path}/integration {posargs}
 
 [testenv:check]
 skip_install=True
